requirements.txt
streamlit>=1.30
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
joblib>=1.3
matplotlib>=3.7
plotly>=5.18
wordcloud>=1.9
PyPDF2>=3.0
beautifulsoup4>=4.12
nltk>=3.8
sentence-transformers
spacy
google-generativeai
sentence-transformers
fpdf2
python-dotenv
en_core_web_sm